Debug CLI modules and logic of workspace setup
* Changed the workspace setup script (`pudl_setup`) to use the user's current
  working directory by default for PUDL_IN or PUDL_OUT if they aren't
  specified.
* Make command line arguments for states case insensitive.
* Integrated module level docstrings into script help messages.
* Update `environment.yml` to include the `defaults` channel again, since there
  are some compilers we need in there.
* Remove development parameters from example settings files that are deployed.
* Fixed accidental clobbering of EPA CEMS per-state data when a new state in
  the same year is downloaded (#382).
* Fixed incorrect calls to `pudl.workspace.setup` modules that omitted the
  `setup` subpackage.

Closes #382

+++ src/pudl/workspace/datastore.py
@@ -570,7 +570,7 @@ def organize(source, year, states, data_dir, unzip=True, dl=True):
     # the data for this source and year... so lets wipe it! Scary!
     destdir = path(source=source, year=year, file=False, data_dir=data_dir)
     if dl:
-        if os.path.exists(destdir):
+        if os.path.exists(destdir) and source != 'epacems':
             shutil.rmtree(destdir)
         # move the new file from wherever it is, to its rightful home.
         if not os.path.exists(destdir):
